+++ hawc/apps/lit/forms.py
@@ -9,8 +9,8 @@ from django import forms
 from django.db import transaction
 from django.db.models import Q
 from django.urls import reverse_lazy
-from ...services.epa import ris
 
+from ...services.utils import ris
 from ..assessment.models import Assessment
 from ..common.forms import BaseFormHelper, addPopupLink
 from . import constants, models
