+++ hawc/services/utils/misc.py
