+++ tests/data/cassettes/test_dsstox/TestDssSubstance.test_bad_dsstox.yaml
@@ -1,0 +1,65 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.23.0
+    method: GET
+    uri: https://actorws.epa.gov/actorws/chemIdentifier/v01/resolve.json?identifier=DTXSID123456789
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAJVSXWvCMBT9L3nwSay6uY+CiCiTolNximNvIbltsyVNlxutIv73pXXtyhDGXkru
+        OT3nnpz2RMbU0pXOiH8i2ogo4JBYEQowxCfj9etLMO50b257d/cPj6RJGEWTED/ZSdkkqYEQjAE+
+        pwpKEI/JBAUnfrs4r49pnarbX0BuD8Xr1cR+pvdRDCpIRnEwhWMJioSLSP9GUQkJWE5Kyydt1E7S
+        GrIFEcXWBWu5aExLKbhIokltewWO6tes0OW1+1ojlAK+jYUFTClzREglQsXMgObiNzAaK845uXzU
+        WuD1Sr5ZFgP7wJ1yn2A+JHnRmgEi8KFzILPFYrpZkryJUD9jVCbJqEncohqicOWWH/MqcmFsbep7
+        HmVWmwxbkNJWpPfl7HFEqw/evt31FJpcOBBVtn7u6HZ+IjX/98tVV8yEohFsVrO/vVwjSjAqvUIy
+        uPwmhU0j63d67UZcPENl+7iPyPn8BdounZvZAgAA
+    headers:
+      Access-Control-Allow-Headers:
+      - x-requested-with
+      Access-Control-Allow-Methods:
+      - POST, GET, OPTIONS, DELETE
+      Access-Control-Allow-Origin:
+      - '*'
+      Allow:
+      - GET, HEAD, POST, PUT, DELETE, TRACE, OPTIONS
+      Cache-Control:
+      - no-cache, no-store, max-age=0
+      Connection:
+      - Keep-Alive
+      Content-Encoding:
+      - gzip
+      Content-Language:
+      - en-US
+      Content-Type:
+      - application/json;charset=UTF-8
+      Date:
+      - Mon, 14 Sep 2020 14:25:00 GMT
+      Expires:
+      - Thu, 01 Jan 1970 00:00:00 GMT
+      Keep-Alive:
+      - timeout=5, max=100
+      Pragma:
+      - no-cache
+      Server:
+      - Apache-Coyote/1.1
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains
+      Transfer-Encoding:
+      - chunked
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - DENY
+    status:
+      code: 200
+      message: OK
+version: 1
