+++ tests/data/cassettes/test_dsstox/TestDssSubstance.test_good_dsstox.yaml
@@ -1,0 +1,132 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.23.0
+    method: GET
+    uri: https://actorws.epa.gov/actorws/chemIdentifier/v01/resolve.json?identifier=1836-75-5
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAJVS72+bMBD9X/hQtdoIP7IsbSRUMTNW1BQ2IGu3qposOMCbjTPbSRpV/d9noInS
+        StW2L2f53r1353d+MAKscMo3xuzB4ILUUQmtIhUBYcwM53T83pxOzInx1iiwFO2L3FJABUJAGWMG
+        GouJEryCVkNy236SpDRmrn02tft7vl12RcjPzDQeSl7tVqr7nm0E+U0WBVO7lxmAYgegKBg/pX+i
+        BljUoia6hK1Gb6L8yyKbz5PsMsg/mIuLMAz95Fvmm11n0pak5v9cLhmhIHXZbWLe3cZv7o695AQ5
+        HkIeOk6QqyOiJ2h/uF1wNJFxGnLBVhR373bciymibpyMB+gaSN0o7dDpu5F9ph3mlJKStPVgXLui
+        9CCJBv9fZD8/38CAKkEYg/K6IQrkEhcaqDCVsEfmgDvydxBc7jGtpGfFSkF5uJYntGig+CVXTD/k
+        qz+Pgn77vAApofQ7b8IkvfLz/GPwQy84HVyu+JWsd1NtsGh104MMk6keZNtZ1Ck0Si1nloULxcVG
+        jmCJRzVf7+5WKaXi99badi0mRUc8J/s5veff4bfE4v9lO9brmoThGhbp/O+S2itGCkytnnI+fNlD
+        taON50zso6aPFVOeXNfG4+MfL/IJtI0DAAA=
+    headers:
+      Access-Control-Allow-Headers:
+      - x-requested-with
+      Access-Control-Allow-Methods:
+      - POST, GET, OPTIONS, DELETE
+      Access-Control-Allow-Origin:
+      - '*'
+      Allow:
+      - GET, HEAD, POST, PUT, DELETE, TRACE, OPTIONS
+      Cache-Control:
+      - no-cache, no-store, max-age=0
+      Connection:
+      - Keep-Alive
+      Content-Encoding:
+      - gzip
+      Content-Language:
+      - en-US
+      Content-Type:
+      - application/json;charset=UTF-8
+      Date:
+      - Mon, 14 Sep 2020 14:25:00 GMT
+      Expires:
+      - Thu, 01 Jan 1970 00:00:00 GMT
+      Keep-Alive:
+      - timeout=5, max=100
+      Pragma:
+      - no-cache
+      Server:
+      - Apache-Coyote/1.1
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains
+      Transfer-Encoding:
+      - chunked
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - DENY
+    status:
+      code: 200
+      message: OK
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.23.0
+    method: GET
+    uri: https://actorws.epa.gov/actorws/chemIdentifier/v01/resolve.json?identifier=DTXSID7020970
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAJVS227bMAz9Fz8ULbbEjrMsbQCjyGRkNeLZXZyg3Yqi0Cza1qZLJik3FP33yXYT
+        pMOuLxTEw3NIHPLRCbHBM7lxRo+OVLSMCAhDCwrKGTnh/DaLwqHnexdDz3nt5FgrYfO98/7bznDQ
+        GdjcUkEBSgFJMAeLJdQoWYCwkN6J95oSZ9Twm/98t6yLwiyby+1DtvqiDRY5PESkrf9je2K2jdyv
+        gHwPoCjsP6e/ogp4JFAVTWFn0dto/nGRxXGaTcP5u87iajKZjNNP2biT2GoqCC3lP5drThloW3aX
+        du7vklf3p0F6hnoBQgE6TZFvI2Jn6PD4dehZIpdsIhVfMWzJqOdfDRHzk7TfQjdAy8pYy87fdL0L
+        a7lkjBIqytZJsWLsKInahfyUvX65khY1inIO5KaiBvQS5xYoMNNwQGLANfkzKKkPmFWys2JjgByv
+        5hnNK8i/6RWv1z5ubkHmoDWQcW1MnKbTxXXjbCE/6HI/yQYrYRsdZbie2ea72paaWBmzHLkuzo1U
+        G92FJe6Wcr3/u0RrI7fu2vNdrlVNvKSH2YKXJ/BdY/X/sjXr95qU4xIWs/jvktYfTnPM3IZy2Z7p
+        sdrJJugNvJOqiQU3gV6XztPTD2aKqzyWAwAA
+    headers:
+      Access-Control-Allow-Headers:
+      - x-requested-with
+      Access-Control-Allow-Methods:
+      - POST, GET, OPTIONS, DELETE
+      Access-Control-Allow-Origin:
+      - '*'
+      Allow:
+      - GET, HEAD, POST, PUT, DELETE, TRACE, OPTIONS
+      Cache-Control:
+      - no-cache, no-store, max-age=0
+      Connection:
+      - Keep-Alive
+      Content-Encoding:
+      - gzip
+      Content-Language:
+      - en-US
+      Content-Type:
+      - application/json;charset=UTF-8
+      Date:
+      - Mon, 14 Sep 2020 14:25:00 GMT
+      Expires:
+      - Thu, 01 Jan 1970 00:00:00 GMT
+      Keep-Alive:
+      - timeout=5, max=100
+      Pragma:
+      - no-cache
+      Server:
+      - Apache-Coyote/1.1
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains
+      Transfer-Encoding:
+      - chunked
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - DENY
+    status:
+      code: 200
+      message: OK
+version: 1
